config: check PATH environment for supported compiler, terminate if not found
With c/config*.py's current behavior, it presumes that the host system uses gcc and it relies on the shell to resolve the location of gcc. This is part of a larger effort to refactor the file to stop relying on the shell.

Instead of doing that, we go through PATH ourselves and only if we find a filename that matches our search parameters (essentially, supporting clang and gcc), we bother executing the compile instructions or we throw an exception.

This commit does cause an intentional bias in favor of clang due to its increased performance and will only choose gcc if clang is not found