Use byte instead of char literals for u8 values
While the differences are confusing, casting a character literal
(which is a 32-bit value) to a u8 is error-prone, since most non-ascii
characters will be truncated to a different value. Byte literals
are validated to be ASCII as compile time so the truncation is safe.

Addresses as clippy::char_lit_as_u8 warning.